Modif Weather

+++ SharpDevelopProjects/Weather/Weather/WeatherData.cs
@@ -47,7 +47,7 @@ namespace WComp.Beans
 	public class Main
 	{
 	    public double temp { get; set; }
-	    public int pressure { get; set; }
+	    public double pressure { get; set; }
 	    public int humidity { get; set; }
 	    public double temp_min { get; set; }
 	    public double temp_max { get; set; }
@@ -56,7 +56,7 @@ namespace WComp.Beans
 	public class Wind
 	{
 	    public double speed { get; set; }
-	    public int deg { get; set; }
+	    public double deg { get; set; }
 	}
 	
 	public class Clouds
